+++ src/test/java/com/sealog/backend/support/base/TestPersistenceBase.java
@@ -1,14 +1,10 @@
 package com.sealog.backend.support.base;
 
-import com.sealog.backend.domain.feature.user.entity.User;
-import com.sealog.backend.domain.feature.user.enums.UserRole;
 import com.sealog.backend.support.extension.ExecutionTimeExtension;
 import com.sealog.backend.support.component.TestDataFactory;
 import com.sealog.backend.support.constant.TestMode;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
